+++ src/dataset/cpp/gigasecond_exercism-felbit/openAI/ChatGPT4_gigasecond_v2.cpp
@@ -2,7 +2,7 @@
 
 namespace gigasecond
 {
-    constexpr int GIGASECOND = 1000000000;
+    constexpr long GIGASECOND = 1000000000;
 
     boost::posix_time::ptime advance(const boost::posix_time::ptime& begin) noexcept
     {
