+++ src/dataset/cpp/nth-prime_exercism-felbit/openAI/ChatGPT4_nth_prime_v2.cpp
@@ -6,7 +6,7 @@ namespace prime {
     int nth(unsigned long n)
     {
         if (n < 1) throw std::domain_error("Non positive number given!");
-        
+
         std::vector<int> primes;
         primes.reserve(n); // Reserve memory to avoid reallocations
         primes.push_back(2); // Start with the first prime number
@@ -16,11 +16,11 @@ namespace prime {
         while (primes.size() < n)
         {
             bool is_prime = true;
-            int limit = static_cast<int>(std::sqrt(candidate)); // Check divisors up to sqrt(candidate)
+            int limit = std::sqrt(candidate);
 
             for (int p : primes)
             {
-                if (p > limit) break; // No need to check beyond sqrt(candidate)
+                if (p > limit) break; // No need to check beyond the square root
                 if (candidate % p == 0)
                 {
                     is_prime = false;
@@ -35,5 +35,5 @@ namespace prime {
 
         return primes.back();
     }
-    
+
 }
